http: fix memsize casting
V220: Suspicious sequence of types castings: memsize -> 32-bit integer -> memsize.

https://pvs-studio.com/en/docs/warnings/v220/

+++ lib/http_aws_sigv4.c
@@ -48,9 +48,9 @@
   do {                                      \
     ret = Curl_hmacit(Curl_HMAC_SHA256,     \
                       (unsigned char *)k,   \
-                      (unsigned int)kl,     \
+                      (size_t)kl,     \
                       (unsigned char *)d,   \
-                      (unsigned int)dl, o); \
+                      (size_t)dl, o); \
     if(ret) {                               \
       goto fail;                            \
     }                                       \
